eeg_pipeline/requirements.txt
kafka-python==2.2.14
pydantic>=2,<3
numpy>=2.3
scipy>=1.16
matplotlib>=3.10
mne>=1.9
packaging>=25.0
tqdm>=4.67
pylsl>=1.17.0
